camera_server/requirements.txt
tornado
pillow
requests